Add more warnings about facts removed
Three reasons for facts being removed from cubes, flow through
suppression, removal of stunted cash flow columns, and incomplete
movements missing an instant fact at the beginning or end.

Flow through suppression is reported in the info messages, but these are
often ignored.  Flow through suppression will now test whether the
removal would remove the last presented location for a fact, and if so,
will not suppress the column.  Also, the statement of equity itself is
not subject to flow through suppression, but it is considered when
testing whether a given fact is reported in "other" reports.

Stunted cash flow columns follows a different logic but also is modified
so that it will not remove the sole occurrence of a fact in the reports.
Note that this will now produce additional columns when the filer has
failed to correctly use the periodStartLabel feature on the beginning
balances for cash.

Duration facts are removed from the statement of changes in equity when
they are not matched with a beginning and ending instant; instant facts
are remove from the statement of changes in equity when they are not
matched with at least one duration.  Previously this was only reported
as INFO, but now, when the removed fact is the only report where it
appears, there is a detailed warning message.

This is currently being tested on the last 100 filings with R9999
reports, and once all have been verified to eliminate the R9999 or
accompany the R9999 with the more informative warning, it will be tested
on all the R9999 cases going back to january 2023.